增加tdengine

+++ internal/model/model.go
@@ -218,10 +218,10 @@ type Alarm struct {
 }
 
 type Trigger struct { //触发条件
-	Type     string  `json:"type"`     //触发条件类型
-	ModelId  string  `json:"modelId"`  //属性ID
-	Operator string  `json:"operator"` //条件
-	Val      float64 `json:"val"`      //数据值
+	Type     string `json:"type"`     //触发条件类型
+	ModelId  string `json:"modelId"`  //属性ID
+	Operator string `json:"operator"` //条件
+	Val      uint64 `json:"val"`      //数据值
 }
 
 type Action struct { //执行动作
